refactor(check): extract verdict and check helpers from check_cmd.py
check_cmd.py was 1514 LOC: 17 helpers stacked above the main `check`
Click command (~820 LOC body), then `replay`. The helpers fell into
two natural groups: verdict-payload computation (~360 LOC), and small
support utilities — tag filtering, judge usage summary, baseline
context formatting (~170 LOC).

Extracted into:
- _check_verdict.py — _compute_check_exit_code, _VerdictOutput,
  _aggregate_cost_delta_ratio, _dedup_recommendations,
  _compute_verdict_payload, _render_verdict_panel,
  _substitute_test_name (372 LOC)
- _check_helpers.py — _all_failures_retry_healed, target/tag
  helpers, _judge_usage_summary, baseline-timestamp helpers (188 LOC)

check_cmd.py is now 1004 LOC: imports + the `check` and `replay`
Click commands. The command bodies themselves are the next-largest
unit; further factoring would require splitting Click options across
modules, which trades clarity for size and isn't worth it here.

Re-exports preserved with noqa: F401 for symbols that tests
monkeypatch via `evalview.commands.check_cmd.<name>`:
_compute_verdict_payload, _all_failures_retry_healed, _cloud_pull,
_load_config_if_exists, plus all extracted private helpers.

Zero functional change. 1792 tests pass (same baseline). Tests that
fail when run in isolation (env-var leak from earlier tests in suite)
are unchanged from before refactor.
